64 bit compat

+++ Source/Categories/NSString+Levenshtein.h
@@ -2,17 +2,19 @@
 //  NSString+Levenshtein.h
 //  PyHelp
 //
-//  Created by Michael Bianco on 12/2/11.
-//  Copyright (c) 2011 MAB Web Design. All rights reserved.
+//  Modified by Michael Bianco on 12/2/11.
+//	<http://mabblog.com>
 //
+//  Created by Rick Bourner on Sat Aug 09 2003.
+//  rick@bourner.com
 
 #import <Foundation/Foundation.h>
 
 @interface NSString (Levenshtein)
 
 // calculate the smallest distance between all words in stringA and stringB
-- (float) compareWithString: (NSString *) stringB matchGain:(int)gain missingCost:(int)cost;
+- (CGFloat) compareWithString: (NSString *) stringB matchGain:(NSInteger)gain missingCost:(NSInteger)cost;
 
 // calculate the distance between two string treating them each as a single word
-- (int) compareWithWord:(NSString *) stringB matchGain:(int)gain missingCost:(int)cost;
+- (NSInteger) compareWithWord:(NSString *) stringB matchGain:(NSInteger)gain missingCost:(NSInteger)cost;
 @end
